+++ lib/MIP/Cli/Mip/Analyse/Rd_dna.pm
@@ -894,6 +894,14 @@ q{Default: grch37_dbsnp_-138-.vcf, grch37_1000g_indels_-phase1-.vcf, grch37_mill
     );
 
     option(
+        q{sv_svdb_query_overlap} => (
+            documentation => q{Database file(s) for annotation},
+            is            => q{rw},
+            isa           => Str,
+        )
+    );
+
+    option(
         q{vcf2cytosure_ar} => (
             cmd_tags      => [q{Analysis recipe switch}],
             documentation =>